Customerのテストの完成
